Moved span generation to component model.
