+++ src/request/component-models/LoggingComponentModel.ts
@@ -2,7 +2,7 @@
 
 import { ComponentModel } from './ComponentModel';
 import { IGlimpse } from '../../IGlimpse';
-import { ILoggingComponentModel, ILoggingLevelModel, ILogMessageModel } from './ILoggingComponentModel';
+import { ILoggingComponentModel, ILoggingLevelModel, ILogMessageModel, ILogMessageSpan } from './ILoggingComponentModel';
 import { ILoggingComponentState } from './ILoggingComponentState';
 import { ILogMessage, ILogMessageReplacedRegion } from '../messages/ILogMessage';
 import { IMessageEnvelope } from '../messages/IMessageEnvelope';
@@ -11,6 +11,8 @@ import { IRequestDetailStore } from '../stores/IRequestDetailStore';
 import _ = require('lodash');
 
 class LogMessageModel implements ILogMessageModel {
+    private _spans: ILogMessageSpan[];
+
     public constructor(private _message: IMessageEnvelope<ILogMessage>, private _ordinal: number) {
     }
 
@@ -32,6 +34,64 @@ class LogMessageModel implements ILogMessageModel {
 
     public get replacedRegions(): ILogMessageReplacedRegion[] {
         return this._message.payload.replacedRegions;
+    }
+
+    public get spans(): ILogMessageSpan[] {
+        if (!this._spans) {
+            this._spans = LogMessageModel.createSpans(this.message, this.replacedRegions);
+        }
+
+        return this._spans;
+    }
+
+    private static createSpans(message: string, replacedRegions: ({ start: number, end: number })[]): ILogMessageSpan[] {
+        message = message || '';
+        replacedRegions = _.sortBy(replacedRegions || [], region => region.start);
+
+        let messageIndex = 0;
+        const messageStructure = [];
+
+        for (let i = 0; i < replacedRegions.length; i++) {
+            const region = replacedRegions[i];
+
+            if (region.start < 0 || region.start >= message.length) {
+                console.warn('The region [%d,%d) exceeds the bounds of the log message (length === %d).', region.start, region.end, message.length);
+
+                continue;
+            }
+
+            if (region.end < 0 || region.end > message.length) {
+                console.warn('The region [%d,%d) exceeds the bounds of the log message (length === %d).', region.start, region.end, message.length);
+
+                continue;
+            }
+
+            if (region.end < region.start) {
+                console.warn('The region [%d,%d) is not a contiguous span in the log message (length === %d).', region.start, region.end, message.length);
+
+                continue;
+            }
+
+            if (region.start < messageIndex) {
+                console.warn('The region [%d,%d) overlaps a previous span in the log message (length === %d).', region.start, region.end, message.length);
+
+                continue;
+            }
+
+            if (messageIndex < region.start) {
+                messageStructure.push({ text: message.substring(messageIndex, region.start) });
+            }
+
+            messageStructure.push({ text: message.substring(region.start, region.end), wasReplaced: true });
+
+            messageIndex = region.end;
+        }
+
+        if (messageIndex < message.length) {
+            messageStructure.push({ text: message.substring(messageIndex, message.length) });
+        }
+
+        return messageStructure;
     }
 }
